backend/requirements.txt
Django
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-more-admin-filters
drf-spectacular
gunicorn
psycopg2-binary
pydub
pyotp
redis
SpeechRecognition
celery
channels
django-jazzmin
whitenoise
